+++ CHANGELOG.md
@@ -6,7 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-## [0.4.1] - 2025-05-01
+## [0.4.5] - 2025-05-01
+### Fixed
+- Filesystem Utils
+- Configuration Reset
+- Certificate Manager Paths
+## [0.4.4] - 2025-05-01
 ### Fixed
 - Docker Menu returning point
 ## [0.4.3] - 2025-04-29